requirements.txt
restkit
biopython
